Fix data types used for list sizes in QEMU capabilities
The QEMU capabilities APIs used a misc of 'int' and
'unsigned int' for variables relating to array sizes.
Change all these to use 'size_t'

+++ src/qemu/qemu_capabilities.h
@@ -180,12 +180,12 @@ virCapsPtr qemuCapsInit(virCapsPtr old_caps);
 int qemuCapsProbeMachineTypes(const char *binary,
                               qemuCapsPtr caps,
                               virCapsGuestMachinePtr **machines,
-                              int *nmachines);
+                              size_t *nmachines);
 
 int qemuCapsProbeCPUModels(const char *qemu,
                            qemuCapsPtr caps,
                            const char *arch,
-                           unsigned int *count,
+                           size_t *count,
                            const char ***cpus);
 
 int qemuCapsExtractVersion(virCapsPtr caps,
